chore(suite): move NetworkSettingsSearchInput file
